docs: update README.md with additional sections on JWT and Refresh Token The README.md file has been updated to include additional sections on JWT and Refresh Token. These sections provide an overview of the logic behind JWT and Refresh Token, and how they are used in the application.
refactor: rename module name in go.mod file
The module name in the go.mod file has been renamed to reflect the new name of the project.

refactor: update import statements in handler and main files
The import statements in the handler and main files have been updated to reflect the new name of the project.

refactor: update import statement in service/user.go file
The import statement in the service/user.go file has been updated to reflect the new name of the project.